[GUI] Move import rides to history FAB
